refactor: integrate ButtonPrimary functionality into Button component
- Integrate ButtonPrimary functionality directly into Button component's primary variant
- Remove separate ButtonPrimary component file and its test
- Update ServicesSection to use Button with variant='primary' instead of ButtonPrimary
- Update demo page to showcase integrated primary button functionality
- Remove ButtonPrimary exports from index.ts
- Update test mocks and comments to reflect the integration
- Maintain all existing functionality: arrow icon, SCSS styling, client-side mounting
